Move audience facet to event wrapper so it can display on both current and past events.
