+++ src/components/LandingPages/Events/Current/index.js
@@ -9,15 +9,8 @@ import Presenter from './presenter'
 
 import { fetchAllEvents } from 'actions/contentful/allEvents'
 import * as statuses from 'constants/APIStatuses'
-import * as helper from 'constants/HelperFunctions'
 
 export class CurrentEventsContainer extends Component {
-  constructor (props) {
-    super(props)
-
-    this.onAudienceFilterApply = this.onAudienceFilterApply.bind(this)
-  }
-
   componentDidMount () {
     const preview = (new URLSearchParams(this.props.location.search)).get('preview') === 'true'
     if (this.props.allEventsStatus === statuses.NOT_FETCHED) {
@@ -25,16 +18,8 @@ export class CurrentEventsContainer extends Component {
     }
   }
 
-  onAudienceFilterApply (selection) {
-    const queryString = helper.buildQueryString(this.props.location.search, 'audience', selection)
-    this.props.history.push(this.props.location.pathname + queryString)
-  }
-
   render () {
-    return <Presenter
-      {...this.props}
-      onAudienceFilterApply={this.onAudienceFilterApply}
-    />
+    return <Presenter {...this.props} />
   }
 }
 
@@ -65,16 +50,6 @@ export const mapStateToProps = (state, ownProps) => {
     return start === dateString || end === dateString || (start < dateString && end >= dateString)
   }) : events
 
-  // Get audience filter
-  const audienceFilter = []
-  const queryParams = ownProps.location.search.replace('?', '').split('&')
-  queryParams.forEach(param => {
-    const split = decodeURIComponent(param).split('=')
-    if (split[0].toLowerCase() === 'audience') {
-      audienceFilter.push(split[1])
-    }
-  })
-
   return {
     pageTitle,
     pageDate,
@@ -82,7 +57,6 @@ export const mapStateToProps = (state, ownProps) => {
     events,
     filteredEvents,
     allEventsStatus: allEvents.status,
-    audienceFilter,
   }
 }
 
@@ -103,10 +77,6 @@ CurrentEventsContainer.propTypes = {
       PropTypes.string,
       PropTypes.object,
     ]),
-    pathname: PropTypes.string,
-  }).isRequired,
-  history: PropTypes.shape({
-    push: PropTypes.func.isRequired,
   }).isRequired,
 }
