+++ src/components/LandingPages/Events/Past/DateFilter/presenter.js
@@ -2,17 +2,18 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Link from 'components/Interactive/Link'
+import styles from './style.module.css'
 
 const DateFilter = (props) => {
   return (
-    <aside aria-label='Select archived month to display' role='navigation' className='dateFilter'>
+    <aside aria-label='Select archived month to display' role='navigation' className={styles.dateFilter}>
       <div className='group'>
         <span className='nav-header'>Archive</span>
-        <ul className='archive'>
+        <ul className={styles.archive}>
           {
             Object.keys(props.eventDates).reverse().map((year) => {
               return (
-                <li key={'filter_' + year} className={props.expanded.includes(year) ? 'expanded' : 'collapsed'}>
+                <li key={'filter_' + year} className={props.expanded.includes(year) ? styles.expanded : styles.collapsed}>
                   { /* eslint-disable-next-line jsx-a11y/anchor-is-valid */ }
                   <a className='yearFilter' onClick={props.yearCallback}>{year}</a>
                   <ul className={'monthFilter' + (props.expanded.includes(year) ? '' : ' hidden')}>
