+++ src/components/LandingPages/Events/Past/presenter.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types'
 
 import DateFilter from './DateFilter'
 import EventsWrapper from '../Wrapper'
-import SideNav from 'components/Layout/Navigation/SideNav'
 
 const Presenter = (props) => {
   return (
@@ -14,12 +13,10 @@ const Presenter = (props) => {
       pageDate={props.pageDate}
       events={props.events}
       filteredEvents={props.filteredEvents}
+      location={props.location}
+      history={props.history}
     >
-      <div className='col-md-4 col-xs-12' style={{ position: 'relative' }}>
-        <SideNav className='column-md'>
-          <DateFilter events={props.events} filterYear={props.filterYear} filterMonth={props.filterMonth} />
-        </SideNav>
-      </div>
+      <DateFilter events={props.events} filterYear={props.filterYear} filterMonth={props.filterMonth} />
     </EventsWrapper>
   )
 }
@@ -31,6 +28,8 @@ Presenter.propTypes = {
   filteredEvents: PropTypes.array.isRequired,
   filterYear: PropTypes.number,
   filterMonth: PropTypes.number,
+  location: PropTypes.object,
+  history: PropTypes.object,
 }
 
 export default Presenter
