+++ src/static/css/global.css
@@ -2347,48 +2347,6 @@ main blockquote p {
   background-position:2px 0 ;
 }
 
-.dateFilter ul {
-  list-style: none;
-  padding-left: 0;
-}
-
-.dateFilter li {
-  list-style: none;
-  padding-left: 10px;
-  position: relative;
-  font-weight: bold;
-}
-
-.dateFilter .archive {
-  margin-top: 8px;
-}
-
-.dateFilter .collapsed::before {
-  color: #b3b3b3;
-  content: "›";
-  font-size: 25px;
-  left: 0;
-  position: absolute;
-  top: -2px;
-  speak: none;
-}
-
-.dateFilter .expanded::before {
-  color: #b3b3b3;
-  content: "›";
-  left: 0px;
-  position: absolute;
-  font-size: 25px;
-  top: -2px;
-  speak: none;
-
-  -webkit-transform: rotate(90deg);
-  -moz-transform: rotate(90deg);
-  -o-transform: rotate(90deg);
-  -ms-transform: rotate(90deg);
-  transform: rotate(90deg);
-}
-
 .yearFilter {
   cursor: pointer;
 }
