+++ src/tests/components/LandingPages/Events/Current/index.test.js
@@ -78,12 +78,6 @@ describe('components/LandingPages/Events/Current', () => {
     it('should not call fetchAllEvents', () => {
       expect(props.fetchAllEvents).not.toHaveBeenCalled()
     })
-
-    it('should update history when filtering audience', () => {
-      const instance = enzymeWrapper.instance()
-      instance.onAudienceFilterApply(['Epic Developer'])
-      expect(enzymeWrapper.props().history.push).toHaveBeenCalled()
-    })
   })
 
   describe('mapStateToProps', () => {
@@ -162,19 +156,6 @@ describe('components/LandingPages/Events/Current', () => {
         filteredEvents: [],
         allEventsStatus: state.allEvents.status,
       }))
-    })
-
-    it('should get audience filter from query string', () => {
-      const state = {
-        allEvents: {
-          status: statuses.SUCCESS,
-          json: testData.allTestEvents,
-        },
-      }
-      const result = mapStateToProps(state, props)
-      const expected = ['you', 'me']
-      expect(result.audienceFilter).toEqual(expect.arrayContaining(expected))
-      expect(result.audienceFilter).toHaveLength(expected.length)
     })
   })
