+++ src/tests/components/LandingPages/Events/Past/presenter.test.js
@@ -57,16 +57,11 @@ describe('components/LandingPages/Events/Past/presenter', () => {
     }))
   })
 
-  it('should render a SideNav component as a child of EventsWrapper', () => {
+  it('should render a DateFilter component as a child of EventsWrapper', () => {
     const wrapper = enzymeWrapper.find(EventsWrapper)
-    expect(wrapper.find(SideNav).exists()).toBe(true)
-  })
+    expect(wrapper.exists()).toBe(true)
 
-  it('should render a DateFilter component as a child of SideNav', () => {
-    const nav = enzymeWrapper.find(SideNav)
-    expect(nav.exists()).toBe(true)
-
-    expect(nav.containsMatchingElement(
+    expect(wrapper.containsMatchingElement(
       <DateFilter events={props.events} filterYear={props.filterYear} filterMonth={props.filterMonth} />
     )).toBe(true)
     // Make sure those props actually had values otherwise we can't verify it is passing anything
